abstract the only parameter
