+++ src/Ushahidi/Modules/V5/DTO/ApikeySearchFields.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Ushahidi\Modules\V5\Helpers\ParameterUtilities;
 
-class ApikeySearchFields
+class ApikeySearchFields extends SearchFields
 {
     /**
      * @var ?string
