+++ src/Ushahidi/Modules/V5/DTO/CSVSearchFields.php
@@ -3,10 +3,9 @@
 namespace Ushahidi\Modules\V5\DTO;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Ushahidi\Modules\V5\Helpers\ParameterUtilities;
 
-class CSVSearchFields
+class CSVSearchFields extends SearchFields
 {
     /**
      * @var array
