+++ src/Ushahidi/Modules/V5/DTO/ContactSearchFields.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Ushahidi\Modules\V5\Helpers\ParameterUtilities;
 
-class ContactSearchFields
+class ContactSearchFields extends SearchFields
 {
     /**
      * @var ?string
