+++ src/Ushahidi/Modules/V5/DTO/DataProvideSearchFields.php
@@ -4,7 +4,7 @@ namespace Ushahidi\Modules\V5\DTO;
 
 use Illuminate\Http\Request;
 
-class DataProvideSearchFields
+class DataProvideSearchFields extends SearchFields
 {
     /**
      * @var ?string
