+++ src/Ushahidi/Modules/V5/DTO/HXLOrganizationSearchFields.php
@@ -3,10 +3,8 @@
 namespace Ushahidi\Modules\V5\DTO;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Ushahidi\Modules\V5\Helpers\ParameterUtilities;
 
-class HXLOrganizationSearchFields
+class HXLOrganizationSearchFields extends SearchFields
 {
     private $q;
     public function __construct(Request $request)
