+++ src/Ushahidi/Modules/V5/DTO/LayerSearchFields.php
@@ -3,10 +3,9 @@
 namespace Ushahidi\Modules\V5\DTO;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Ushahidi\Modules\V5\Helpers\ParameterUtilities;
 
-class LayerSearchFields
+class LayerSearchFields extends SearchFields
 {
     
     private $type;
