+++ src/Ushahidi/Modules/V5/DTO/SearchFields.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Ushahidi\Modules\V5\DTO;
+
+use Illuminate\Http\Request;
+
+abstract class SearchFields
+{
+    
+}
